refactor: 예외 테스트를 service 계층 테스트에서 진행
