+++ src/test/java/org/greedy/ddarahang/api/service/TravelCourseServiceTest.java
@@ -3,16 +3,25 @@ package org.greedy.ddarahang.api.service;
 import jakarta.transaction.Transactional;
 import org.greedy.ddarahang.api.dto.TravelCourseListResponse;
 import org.greedy.ddarahang.api.dto.TravelCourseResponse;
+import org.greedy.ddarahang.common.exception.InvalidCountryNameException;
+import org.greedy.ddarahang.common.exception.MissingIdException;
+import org.greedy.ddarahang.common.exception.NotFoundTravelCourseDetailException;
 import org.greedy.ddarahang.common.fixture.CountryFixture;
+import org.greedy.ddarahang.common.fixture.PlaceFixture;
 import org.greedy.ddarahang.common.fixture.RegionFixture;
+import org.greedy.ddarahang.common.fixture.TravelCourseDetailFixture;
 import org.greedy.ddarahang.common.fixture.TravelCourseFixture;
 import org.greedy.ddarahang.common.fixture.VideoFixture;
 import org.greedy.ddarahang.db.country.Country;
 import org.greedy.ddarahang.db.country.CountryRepository;
+import org.greedy.ddarahang.db.place.Place;
+import org.greedy.ddarahang.db.place.PlaceRepository;
 import org.greedy.ddarahang.db.region.Region;
 import org.greedy.ddarahang.db.region.RegionRepository;
 import org.greedy.ddarahang.db.travelCourse.TravelCourse;
 import org.greedy.ddarahang.db.travelCourse.TravelCourseRepository;
+import org.greedy.ddarahang.db.travelCourseDetail.TravelCourseDetail;
+import org.greedy.ddarahang.db.travelCourseDetail.TravelCourseDetailRepository;
 import org.greedy.ddarahang.db.video.Video;
 import org.greedy.ddarahang.db.video.VideoRepository;
 import org.junit.jupiter.api.BeforeEach;
@@ -28,11 +37,15 @@ import java.time.LocalDate;
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 
 @ExtendWith(MockitoExtension.class)
 @AutoConfigureMockMvc
 @SpringBootTest
 class TravelCourseServiceTest {
+
+    @Autowired
+    private TravelCourseService travelCourseService;
 
     @Autowired
     private CountryRepository countryRepository;
@@ -41,34 +54,43 @@ class TravelCourseServiceTest {
     private RegionRepository regionRepository;
 
     @Autowired
+    private PlaceRepository placeRepository;
+
+    @Autowired
     private VideoRepository videoRepository;
 
     @Autowired
     private TravelCourseRepository travelCourseRepository;
 
-
     @Autowired
-    private TravelCourseService travelCourseService;
+    private TravelCourseDetailRepository travelCourseDetailRepository;
 
     private Country country;
     private Region region;
+    private Place place;
     private Video video;
     private TravelCourse travelCourse;
+    private TravelCourseDetail travelCourseDetail;
 
     @BeforeEach
     void setUp() {
+        travelCourseDetailRepository.deleteAll();
         travelCourseRepository.deleteAll();
         videoRepository.deleteAll();
+        placeRepository.deleteAll();
         regionRepository.deleteAll();
         countryRepository.deleteAll();
+
         prepareTestData();
     }
 
     private void prepareTestData() {
         country = countryRepository.save(CountryFixture.getMockCountry());
-        region = regionRepository.save(RegionFixture.getMockRegion(country));
-        video = videoRepository.save(VideoFixture.getMockVideo(LocalDate.now()));
+        region = regionRepository.save(RegionFixture.getMockRegion_1(country));
+        place = placeRepository.save(PlaceFixture.getMockPlace_1(region));
+        video = videoRepository.save(VideoFixture.getMockVideo_1(LocalDate.now()));
         travelCourse = travelCourseRepository.save(TravelCourseFixture.getMockTravelCourse(video, country, region));
+        travelCourseDetail = travelCourseDetailRepository.save(TravelCourseDetailFixture.getMockTravelCourseDetail(travelCourse,place));
     }
 
     /**
@@ -77,8 +99,9 @@ class TravelCourseServiceTest {
     @Nested
     @Transactional
     class 여행_목록_조회_메서드 {
+
         @Test
-        void regionName이_없는_경우() {
+        void countryName과_regionName이_모두_있으면_데이터가_정상적으로_반환된다() {
             //Given & When
             List<TravelCourseListResponse> responses = travelCourseService.getTravelCourses(country.getName(), region.getName());
 
@@ -88,13 +111,29 @@ class TravelCourseServiceTest {
         }
 
         @Test
-        void regionName이_있는_경우() {
+        void countryName은_있고_regionName이_없어도_데이터가_정상적으로_반환된다() {
             //Given & When
             List<TravelCourseListResponse> responses = travelCourseService.getTravelCourses(country.getName(), region.getName());
 
             //Then
             assertThat(responses.get(0).creator()).isEqualTo(video.getCreator());
             assertThat(responses.get(0).thumbnailUrl()).isEqualTo(video.getThumbnailUrl());
+        }
+
+        @Test
+        void countryName이_null이면_InvalidCountryNameException이_터진다() {
+            // Given & When
+            TravelCourseService service = new TravelCourseService(travelCourseRepository, travelCourseDetailRepository);
+
+            assertThrows(InvalidCountryNameException.class, () -> service.getTravelCourses(null, region.getName()));
+        }
+
+        @Test
+        void countryName이_비어있으면_InvalidCountryNameException이_터진다() {
+            // Given & When
+            TravelCourseService service = new TravelCourseService(travelCourseRepository, travelCourseDetailRepository);
+
+            assertThrows(InvalidCountryNameException.class, () -> service.getTravelCourses("", region.getName()));
         }
     }
 
@@ -107,65 +146,26 @@ class TravelCourseServiceTest {
     class 여행_상세_조회_메서드 {
 
         @Test
-        void 존재하는_id를_조회하면_정상적으로_반환되며_개수가_일치한다() {
+        void 존재하는_id를_조회하면_정상적으로_반환된다() {
             // When
             TravelCourseResponse response = travelCourseService.getTravelCourseDetail(travelCourse.getId());
 
             // Then
             assertThat(response.creator()).isEqualTo(travelCourse.getVideo().getCreator());
         }
-    }
-
-
-    /**
-     * 🔹 업로드 날짜 내림차순 정렬 (`getSortedByUploadDate`)
-     */
-    @Nested
-    @Transactional
-    class 업로드_날짜_내림차순_정렬_메서드 {
 
         @Test
-        void 업로드_날짜가_최신순으로_정렬될때_반환_개수가_일치한다() {
-            // Given & When
-            List<TravelCourseListResponse> responses = travelCourseService.getSortedByUploadDate(country.getName(), region.getName());
+        void 여행_상세_조회_아이디가_null이면_MissingIdException_발생() {
+            TravelCourseService service = new TravelCourseService(travelCourseRepository, travelCourseDetailRepository);
 
-            // Then
-            assertThat(responses.size()).isEqualTo(1);
+            assertThrows(MissingIdException.class, () -> service.getTravelCourseDetail(null));
         }
 
         @Test
-        void regionName이_없는_경우에도_정상_조회되며_반환_개수가_일치한다() {
-            // Given & When
-            List<TravelCourseListResponse> responses = travelCourseService.getSortedByUploadDate(country.getName(), "");
+        void 해당_id를_가진_데이터가_없으면_NotFoundTravelCourseDetailException_발생() {
+            TravelCourseService service = new TravelCourseService(travelCourseRepository, travelCourseDetailRepository);
 
-            // Then
-            assertThat(responses.size()).isEqualTo(1);
-        }
-    }
-
-    /**
-     * 🔹 조회수 내림차순 정렬 (`getSortedByViewCount`)
-     */
-    @Nested
-    @Transactional
-    class 조회수_내림차순_정렬_메서드 {
-
-        @Test
-        void 조회수가_높은순으로_정렬될때_반환_개수가_일치한다() {
-            // Given & When
-            List<TravelCourseListResponse> responses = travelCourseService.getSortedByViewCount(country.getName(), region.getName());
-
-            // Then
-            assertThat(responses.size()).isEqualTo(1);
-        }
-
-        @Test
-        void regionName이_없는_경우에도_정상_조회되며_반환_개수가_일치한다() {
-            // Given & When
-            List<TravelCourseListResponse> responses = travelCourseService.getSortedByViewCount(country.getName(), "");
-
-            // Then
-            assertThat(responses.size()).isEqualTo(1);
+            assertThrows(NotFoundTravelCourseDetailException.class, () -> service.getTravelCourseDetail(-1L));
         }
     }
 }
